+++ wp-admin/includes/plugin-install.php
@@ -21,8 +21,22 @@
  * @return string $button The markup for the dependency row button.
  */
 function wp_get_plugin_action_button( $name, $data, $compatible_php, $compatible_wp ) {
-	$button = '';
-	$status = install_plugin_install_status( $data );
+	$button                   = '';
+	$status                   = install_plugin_install_status( $data );
+	$plugin_dependency_met    = true;
+	$data['requires_plugins'] = isset( $data['requires_plugins'] ) ? $data['requires_plugins'] : array();
+
+	// Check if plugin dependency is installed and active.
+	foreach ( $data['requires_plugins'] as $dependency ) {
+		$plugin_dependency_met = false;
+		$active_plugins        = get_option( 'active_plugins' );
+		foreach ( $active_plugins as $plugin_file ) {
+			if ( str_contains( $plugin_file, '/' ) && explode( '/', $plugin_file )[0] === $dependency ) {
+				$plugin_dependency_met = true;
+				break;
+			}
+		}
+	}
 
 	sprintf(
 		'<a class="install-now button" data-slug="%s" href="%s" aria-label="%s" data-name="%s">%s</a>',
@@ -38,7 +52,7 @@ function wp_get_plugin_action_button( $name, $data, $compatible_php, $compatible
 		switch ( $status['status'] ) {
 			case 'install':
 				if ( $status['url'] ) {
-					if ( $compatible_php && $compatible_wp ) {
+					if ( $compatible_php && $compatible_wp && $plugin_dependency_met ) {
 						$button = sprintf(
 							'<a class="install-now button" data-slug="%s" href="%s" aria-label="%s" data-name="%s">%s</a>',
 							esc_attr( $data['slug'] ),