+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* change plugin card button to 'Cannot Install' if dependencies not met
 
 #### 1.11.0 / 2023-05-21
 * add **Requires:** data to plugin cards of uninstalled plugins where repo plugins have `Requires Plugins` header set